+++ Changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added unstable `algorithms::div` module and improved div algorithm.
+
 ## [1.4.1] — 2022-10-15
 
 ### Changed